♻️ Refactor: 랜딩 페이지 로직 분리
