+++ src/components/landing/index.ts
@@ -1,0 +1,3 @@
+export { default as ServiceTitle } from './ServiceTitle';
+export { default as ServiceLogo } from './ServiceLogo';
+export { default as KakaoLoginButton } from './KakaoLoginButton';
